refactor(inline_test_footprint)!: drop layout checks, rename from unit_test_file_layout
The rule conflated two concerns. Its purpose is to keep inline
unit-test code from growing too large in a production file (and,
optionally, to forbid it outright). The on-disk external-layout
machinery bolted onto it served nobody, so remove it:

- `external_layout` (`nested` / `sibling` / `any`) and the whole
  canonical-location check on external `#[cfg(test)] mod X;`.
- `flag_unexpected_sibling`, the stray-migration-file check.
- `test_module_names`, the inline-footprint name filter.

What remains is the inline footprint: `inline_style`,
`inline_max_lines`, and `inline_max_fraction_of_file`. An external
test module is now simply neutral. The extraction-target help text
keeps suggesting the canonical nested path.

With the layout axis gone, "file layout" no longer describes the
rule, so rename `unit_test_file_layout` to `inline_test_footprint`,
which names exactly the quantity it now measures.

Resolves https://github.com/KSXGitHub/perfectionist/issues/196.
