+++ rules/inline_test_footprint.md
@@ -1,0 +1,104 @@
+<!-- Generated by `gen-docs write-md` from `src/rules/`. Do not edit by hand; run `just gen-rules-md` after changing a rule source. -->
+
+# `perfectionist::inline_test_footprint`
+
+**Default state:** `active`  
+**Source:** [`src/rules/inline_test_footprint.rs`](../src/rules/inline_test_footprint.rs)
+
+> inline test code should be extracted to a separate file
+
+## What it does
+
+Caps how much inline unit-test code a production file may carry.
+Inline test code — `#[cfg(test)] mod X { ... }` blocks, `#[test]
+fn`s, `#[cfg(test)] fn` helpers, and any other `#[cfg(test)]`
+item — is summed per file. The default `external_when_long` style
+flags a file once its inline-test footprint crosses
+`inline_max_lines` (or the optional
+`inline_max_fraction_of_file`); `external_only` flags every inline
+test item regardless of length. A file whose top-level items are
+*entirely* test code is exempt — it is itself a valid extraction
+target.
+
+An external `#[cfg(test)] mod <name>;` (its body in a separate
+file) is neutral: it is already extracted, so it neither charges
+the footprint nor is otherwise checked.
+
+Only the library or binary crate is checked. Integration tests
+(`tests/`), benchmarks (`benches/`), and examples (`examples/`)
+are separate targets, not the library or binary whose unit-test
+footprint this rule governs; for those compiled under `cfg(test)`
+their top-level `#[test]` functions *are* the target rather than
+unit tests misplaced in a production file, so they are left
+untouched.
+
+## Why restrict this?
+
+This is a stylistic preference, not a correctness issue. Both
+source projects keep large test suites out of the production
+file, so the file an editor tab, a `grep` hit, or a diff shows
+is production code rather than a wall of fixtures. The threshold
+is deliberately configurable because the exact budget varies by
+project.
+
+## Example
+
+**Avoid:**
+
+```rust,ignore
+// File: foo.rs
+#[cfg(test)]
+mod tests {
+    /* ... 200 lines of test code ... */
+}
+```
+
+**Prefer:**
+
+```rust,ignore
+// File: foo.rs
+mod tests;
+```
+
+```rust,ignore
+// File: foo/tests.rs
+/* ... 200 lines of test code ... */
+```
+
+## Configuration
+
+Configure via `dylint.toml` under `["perfectionist::inline_test_footprint"]`. Every field is optional; the per-field prose below states the default.
+
+### `inline_style`: `InlineStyle` (optional)
+
+How inline test modules are handled. Defaults to
+`external_when_long`.
+
+### `inline_max_lines`: `unsigned integer` (optional)
+
+Absolute cap, in lines, on the summed inline-test footprint of a
+file under `external_when_long`. Always active.
+
+### `inline_max_fraction_of_file`: `float` (optional)
+
+Optional relative cap: the share `inline_test_lines / file_lines`
+a file's inline tests may occupy under `external_when_long`.
+Accepted values are `0.0 <= x < 1.0`; omit the key to disable the
+relative cap (the default).
+
+### Types
+
+#### `InlineStyle` (enum)
+
+How inline test code is treated (the `inline_style` knob).
+
+##### `"external_only"` (Rust: `ExternalOnly`)
+
+Every inline test item is flagged; all test code must move to an
+external `mod <name>;`. Matches pacquet's strict policy.
+
+##### `"external_when_long"` (Rust: `ExternalWhenLong`)
+
+Inline test code is allowed up to the configured budget; beyond
+that it must move to a file. Matches parallel-disk-usage's
+guidance. The default.
