--- rules/unit_test_file_layout.md
@@ -1,173 +1,0 @@
-<!-- Generated by `gen-docs write-md` from `src/rules/`. Do not edit by hand; run `just gen-rules-md` after changing a rule source. -->
-
-# `perfectionist::unit_test_file_layout`
-
-**Default state:** `active`  
-**Source:** [`src/rules/unit_test_file_layout.rs`](../src/rules/unit_test_file_layout.rs)
-
-> unit-test code is in the wrong file or exceeds the inline-test budget
-
-## What it does
-
-Enforces where a crate's unit-test code lives. Two independent
-axes are checked:
-
-1. **External-file layout.** An external
-   `#[cfg(test)] mod <name>;` must resolve to the canonical
-   on-disk location. By default that is the nested
-   `<parent>/<name>.rs` form (tests of `src/foo.rs` live in
-   `src/foo/tests.rs`); for such a file the flattened sibling
-   `src/foo_tests.rs` and the skipped-intermediate `src/tests.rs`
-   are flagged. A directory-owning parent (`lib.rs` / `main.rs` /
-   `mod.rs`) is the exception: its children already live beside
-   it, so `mod tests;` in `src/lib.rs` canonically resolves to
-   `src/tests.rs` (not `src/lib/tests.rs`) and is *not* flagged —
-   matching where Cargo loads it. The `sibling` style also
-   accepts the flattened form; `any` skips the layout check.
-2. **Inline footprint.** Inline test code — `#[cfg(test)] mod X
-   { ... }` blocks, `#[test] fn`s, `#[cfg(test)] fn` helpers,
-   and any other `#[cfg(test)]` item — is summed per file. The
-   default `external_when_long` style flags a file once its
-   inline-test footprint crosses `inline_max_lines` (or the
-   optional `inline_max_fraction_of_file`); `external_only`
-   flags every inline test item regardless of length. A file
-   whose top-level items are *entirely* test code is exempt — it
-   is itself a valid extraction target.
-
-The module identifier is irrelevant to the layout rule; only the
-file's position relative to its parent matters.
-
-Only the library or binary crate is checked. Integration tests
-(`tests/`), benchmarks (`benches/`), and examples (`examples/`)
-are separate targets, not the library or binary whose unit-test
-layout this rule governs; for those compiled under `cfg(test)`
-their top-level `#[test]` functions *are* the target rather than
-unit tests misplaced in a production file, so they are left
-untouched.
-
-## Why restrict this?
-
-This is a stylistic preference, not a correctness issue. Both
-source projects keep large test suites out of the production
-file, so the file an editor tab, a `grep` hit, or a diff shows
-is production code rather than a wall of fixtures; and they put
-the extracted file in a predictable place so a reader always
-knows where a module's tests are. The thresholds and the
-nested-vs-sibling choice are deliberately configurable because
-the exact budget and directory shape vary by project.
-
-## Example
-
-**Avoid:**
-
-```rust,ignore
-// File: foo.rs
-#[cfg(test)]
-mod tests {
-    /* ... 200 lines of test code ... */
-}
-```
-
-**Prefer:**
-
-```rust,ignore
-// File: foo.rs
-mod tests;
-```
-
-```rust,ignore
-// File: foo/tests.rs
-/* ... 200 lines of test code ... */
-```
-
-## Configuration
-
-Configure via `dylint.toml` under `["perfectionist::unit_test_file_layout"]`. Every field is optional; the per-field prose below states the default.
-
-### `inline_style`: `InlineStyle` (optional)
-
-How inline test modules are handled. Defaults to
-`external_when_long`.
-
-### `inline_max_lines`: `unsigned integer` (optional)
-
-Absolute cap, in lines, on the summed inline-test footprint of a
-file under `external_when_long`. Always active.
-
-### `inline_max_fraction_of_file`: `float` (optional)
-
-Optional relative cap: the share `inline_test_lines / file_lines`
-a file's inline tests may occupy under `external_when_long`.
-Accepted values are `0.0 <= x < 1.0`; omit the key to disable the
-relative cap (the default).
-
-### `external_layout`: `ExternalLayout` (optional)
-
-How external test files must be laid out on disk. Defaults to
-`nested`.
-
-### `flag_unexpected_sibling`: `boolean` (optional)
-
-Under `nested`, also flag a flattened `<parent>_<name>.rs` sibling
-left on disk for a module whose nested file already exists.
-Defaults to true.
-
-### `test_module_names`: `[string]` (optional)
-
-Module names the inline-style footprint is scoped to. Empty (the
-default) counts every inline test item — `#[cfg(test)] mod`
-blocks of any name, `#[test] fn`s, and other `#[cfg(test)]`
-items. When non-empty, the budget is measured *only* over
-`#[cfg(test)] mod <name>` blocks whose `<name>` is listed; bare
-top-level test items (which have no module name) are then out of
-scope. Set this when a project keeps its inline tests in named
-modules and wants the budget to track those specifically.
-
-### Types
-
-#### `InlineStyle` (enum)
-
-How inline test code is treated (the `inline_style` knob).
-
-##### `"external_only"` (Rust: `ExternalOnly`)
-
-Every inline test item is flagged; all test code must move to an
-external `mod <name>;`. Matches pacquet's strict policy.
-
-##### `"external_when_long"` (Rust: `ExternalWhenLong`)
-
-Inline test code is allowed up to the configured budget; beyond
-that it must move to a file. Matches parallel-disk-usage's
-guidance. The default.
-
-#### `ExternalLayout` (enum)
-
-How external test files must be laid out on disk (the
-`external_layout` knob).
-
-##### `"nested"` (Rust: `Nested`)
-
-`src/foo.rs`'s `mod bar;` must resolve to `src/foo/bar.rs`.
-
-```text
-src/
-├── lib.rs
-├── foo.rs         declares  #[cfg(test)] mod tests;
-└── foo/
-    └── tests.rs   holds the test code
-```
-
-##### `"sibling"` (Rust: `Sibling`)
-
-Also accept the flattened `src/foo_bar.rs` form.
-
-```text
-src/
-├── lib.rs
-├── foo.rs         declares  #[cfg(test)] mod tests;
-└── foo_tests.rs   holds the test code
-``````
-
-##### `"any"` (Rust: `Any`)
-
-Accept whichever path Cargo loads; skip the layout check.
