+++ src/rules.rs
@@ -5,6 +5,7 @@ pub mod derive_ordering;
 pub mod flat_module_pattern;
 pub mod import_granularity;
 pub mod import_grouping;
+pub mod inline_test_footprint;
 pub mod lint_reason_from_comment;
 pub mod lint_silence_reason;
 pub mod macro_argument_binding;
@@ -25,5 +26,4 @@ pub mod single_letter_static_item;
 pub mod unicode_ellipsis_in_comments;
 pub mod unicode_ellipsis_in_docs;
 pub mod unicode_ellipsis_in_panic_messages;
-pub mod unit_test_file_layout;
 pub mod unknown_perfectionist_lints;
